Seperated authentication functionality from user routes | Implemented authentication middleware
