+++ routes/user.js
@@ -2,10 +2,6 @@ const express = require('express');
 const router = express.Router();
 const UserController = require('../controllers/user')
 
-router.post('/register', UserController.register_user);
-
-router.post('/login', UserController.login_user);
-
 router.get('/', UserController.fetchUsers);
 
 router.get('/filter/:userId', UserController.fetchUsersFilter);
